Refactor: JwtPair를 record로 수정
설명:
- 코드 가독성 향상과 불변성 유지를 위해 JwtPair를 record로 수정하려 합니다.
- 또, JwtPair는 DTO로써 이용되기 때문에 dto 디렉토리로 옮기려 합니다.

작업:
- JwtPair를 record로 수정
- JwtPair를 하위 dto폴더 이동
