--- src/main/java/com/gamsung/backend/global/jwt/JwtPair.java
@@ -1,19 +1,0 @@
-package com.gamsung.backend.global.jwt;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class JwtPair {
-    private String accessToken;
-    private String refreshToken;
-
-    @Builder
-    private JwtPair(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-}
