+++ src/main/java/com/gamsung/backend/global/jwt/controller/JwtController.java
@@ -1,8 +1,8 @@
 package com.gamsung.backend.global.jwt.controller;
 
 import com.gamsung.backend.global.common.ApiResponse;
-import com.gamsung.backend.global.jwt.JwtPair;
 import com.gamsung.backend.global.jwt.controller.request.RefreshAccessTokenRequest;
+import com.gamsung.backend.global.jwt.dto.JwtPair;
 import com.gamsung.backend.global.jwt.service.JwtService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
