+++ src/main/java/com/gamsung/backend/global/jwt/dto/JwtPair.java
@@ -1,0 +1,10 @@
+package com.gamsung.backend.global.jwt.dto;
+
+public record JwtPair(
+        String accessToken,
+        String refreshToken
+) {
+    public static JwtPair from(String accessToken, String refreshToken) {
+        return new JwtPair(accessToken, refreshToken);
+    }
+}
