+++ src/main/java/com/gamsung/backend/global/jwt/service/JwtService.java
@@ -1,8 +1,8 @@
 package com.gamsung.backend.global.jwt.service;
 
-import com.gamsung.backend.global.jwt.JwtPair;
 import com.gamsung.backend.global.jwt.JwtProvider;
 import com.gamsung.backend.global.jwt.controller.request.RefreshAccessTokenRequest;
+import com.gamsung.backend.global.jwt.dto.JwtPair;
 import com.gamsung.backend.global.jwt.dto.JwtPayload;
 import com.gamsung.backend.global.jwt.entity.JwtBlackListRedisEntity;
 import com.gamsung.backend.global.jwt.entity.JwtRedisEntity;
@@ -38,10 +38,7 @@ public class JwtService {
                 .expiration(refreshExpiration)
                 .build());
 
-        return JwtPair.builder()
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
+        return JwtPair.from(accessToken, refreshToken);
     }
 
     public JwtPayload verifyAccessToken(String jwtAccessToken) {
@@ -75,10 +72,7 @@ public class JwtService {
 
         String newAccessToken = jwtProvider.createToken(newJwtPayload, accessExpiration);
 
-        return JwtPair.builder()
-                .accessToken(newAccessToken)
-                .refreshToken(request.refreshToken())
-                .build();
+        return JwtPair.from(newAccessToken, request.refreshToken());
     }
 
     public void deleteRefreshTokenAndAddAccessTokenToBlackList(String email, String accessToken) {
