+++ src/test/java/com/gamsung/backend/domain/member/service/MemberServiceTest.java
@@ -9,7 +9,7 @@ import com.gamsung.backend.domain.member.dto.response.MemberRegisterResponse;
 import com.gamsung.backend.domain.member.entity.Member;
 import com.gamsung.backend.domain.member.repository.MemberRepository;
 import com.gamsung.backend.global.factory.MemberTestFactory;
-import com.gamsung.backend.global.jwt.JwtPair;
+import com.gamsung.backend.global.jwt.dto.JwtPair;
 import com.gamsung.backend.global.jwt.dto.JwtPayload;
 import com.gamsung.backend.global.jwt.service.JwtService;
 import org.junit.jupiter.api.Assertions;
@@ -58,10 +58,8 @@ class MemberServiceTest {
         public void successToLoginMember() {
             // given
             Member member = MemberTestFactory.createMemberWithRandomValues(false);
-            JwtPair testToken = JwtPair.builder()
-                    .accessToken("fake-access-token")
-                    .refreshToken("fake-refresh-token")
-                    .build();
+            JwtPair testToken = JwtPair.from("fake-access-token", "fake-refresh-token");
+
             MemberLoginRequest memberLoginRequest = new MemberLoginRequest(
                     member.getEmail(), member.getPassword());
 
@@ -125,15 +123,12 @@ class MemberServiceTest {
         public void successToLogoutMember() {
             // given
             Member member = MemberTestFactory.createMemberWithRandomValues(false);
-            JwtPair testToken = JwtPair.builder()
-                    .accessToken("fake-access-token")
-                    .refreshToken("fake-refresh-token")
-                    .build();
+            JwtPair testToken = JwtPair.from("fake-access-token", "fake-refresh-token");
 
             doNothing().when(jwtService).deleteRefreshTokenAndAddAccessTokenToBlackList(anyString(), anyString());
 
             // when
-            MemberLogoutResponse memberLogoutResponse = memberService.logout(member.getEmail(), testToken.getAccessToken());
+            MemberLogoutResponse memberLogoutResponse = memberService.logout(member.getEmail(), testToken.accessToken());
 
             // then
             Assertions.assertEquals(memberLogoutResponse.message(), "로그아웃에 성공했습니다.");
