+++ src/test/java/com/gamsung/backend/domain/order/controller/OrderControllerIntegrationTest.java
@@ -7,7 +7,7 @@ import com.gamsung.backend.domain.member.entity.Member;
 import com.gamsung.backend.domain.member.service.MemberService;
 import com.gamsung.backend.global.common.BaseRedisContainerTest;
 import com.gamsung.backend.global.factory.MemberTestFactory;
-import com.gamsung.backend.global.jwt.JwtPair;
+import com.gamsung.backend.global.jwt.dto.JwtPair;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -39,10 +39,7 @@ public class OrderControllerIntegrationTest extends BaseRedisContainerTest {
         MemberLoginResponse memberLoginResponse = memberService.login(
                 new MemberLoginRequest(member.getEmail(), member.getPassword())
         );
-        jwtPair = JwtPair.builder()
-                .accessToken(memberLoginResponse.accessToken())
-                .refreshToken(memberLoginResponse.refreshToken())
-                .build();
+        jwtPair = JwtPair.from(memberLoginResponse.accessToken(), memberLoginResponse.refreshToken());
     }
 
     @Test
@@ -53,7 +50,7 @@ public class OrderControllerIntegrationTest extends BaseRedisContainerTest {
         //given
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.set(HttpHeaders.AUTHORIZATION, JWT_TOKEN_PREFIX + jwtPair.getAccessToken());
+        headers.set(HttpHeaders.AUTHORIZATION, JWT_TOKEN_PREFIX + jwtPair.accessToken());
 
         String jsonPayload = """
                 [
@@ -99,7 +96,7 @@ public class OrderControllerIntegrationTest extends BaseRedisContainerTest {
         //given
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.set(HttpHeaders.AUTHORIZATION, JWT_TOKEN_PREFIX + jwtPair.getAccessToken());
+        headers.set(HttpHeaders.AUTHORIZATION, JWT_TOKEN_PREFIX + jwtPair.accessToken());
 
         // Act & Assert
         mockMvc.perform(MockMvcRequestBuilders.get("/v1/order/check")
@@ -122,7 +119,7 @@ public class OrderControllerIntegrationTest extends BaseRedisContainerTest {
         //given
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.set(HttpHeaders.AUTHORIZATION, JWT_TOKEN_PREFIX + jwtPair.getAccessToken());
+        headers.set(HttpHeaders.AUTHORIZATION, JWT_TOKEN_PREFIX + jwtPair.accessToken());
 
         // Act & Assert
         mockMvc.perform(MockMvcRequestBuilders.get("/v1/order/me")
